requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   openai
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==46.0.3
    # via google-auth
distro==1.9.0
    # via
    #   google-genai
    #   openai
google-api-core==2.29.0
    # via google-api-python-client
google-api-python-client==2.188.0
    # via jobtrace (pyproject.toml)
google-auth==2.48.0
    # via
    #   jobtrace (pyproject.toml)
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-genai==1.60.0
    # via jobtrace (pyproject.toml)
googleapis-common-protos==1.72.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.12.0
    # via openai
keyboard==0.13.5
    # via jobtrace (pyproject.toml)
mss==10.1.0
    # via jobtrace (pyproject.toml)
openai==2.15.0
    # via jobtrace (pyproject.toml)
pillow==12.1.0
    # via jobtrace (pyproject.toml)
proto-plus==1.27.0
    # via google-api-core
protobuf==6.33.4
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==7.2.1
    # via jobtrace (pyproject.toml)
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   jobtrace (pyproject.toml)
    #   google-genai
    #   openai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.2
    # via httplib2
pyside6==6.10.1
    # via jobtrace (pyproject.toml)
pyside6-addons==6.10.1
    # via pyside6
pyside6-essentials==6.10.1
    # via
    #   pyside6
    #   pyside6-addons
python-dotenv==1.2.1
    # via jobtrace (pyproject.toml)
pywin32==311
    # via jobtrace (pyproject.toml)
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-genai
rsa==4.9.1
    # via google-auth
shiboken6==6.10.1
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
tenacity==9.1.2
    # via google-genai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
websockets==15.0.1
    # via google-genai
